ti: Add more packages and make weston the OOB
Default and officially supported TI images will be the console image
with weston installed and running on top of it.

Therefore add weston in the list of packages to install. Also add a
systemd service file to launch weston by-default on boot.

Further, add `udhcpc` tool in list of packages to install if the board
supports M2-CC33x1 card. This tool is required to configure the device.

Note: Weston does not launch on first boot, since the booting process
kills it. It does launch on second boot onwards. Currently, full-fledged
DEs make the armbian-firstlogin script launch them after the script
finishes setup. But there is no logic yet to start other systemd
services such as weston. Such logic should be added later.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>