extensions: ti-debpkgs: Call `dkms autoinstall` after kernel install
Currently the only dkms package is ti-img-rogue-driver-<platform>-dkms.
There could be more in the future.

Since custom packages are installed before Kernel headers are, DKMS
doesn't bind the module to the kernel during the build. As a result,
this occurs during the first boot, and takes close to 5 minutes.

So fix this by calling `dkms autoinstall` after the kernel headers are
installed. This command binds all out-of-tree kernel modules to the
installed kernel, and the first boot thus finishes normally.

Signed-off-by: Suhaas Joshi <s-joshi@ti.com>